+++ src/main/kotlin/racing/domain/exception/CarNameLengthExcessException.kt
@@ -1,3 +1,3 @@
-package racing.exception
+package racing.domain.exception
 
 class CarNameLengthExcessException(override val message: String) : RuntimeException(message)
