+++ src/main/kotlin/racing/domain/race/CarRace.kt
@@ -1,6 +1,6 @@
-package racing.race
+package racing.domain.race
 
-import racing.common.CarRaceProperty
+import racing.domain.common.CarRaceProperty
 import racing.view.ResultView
 
 class CarRace(
