+++ src/main/kotlin/racing/domain/utils/RandomNumberGenerator.kt
@@ -1,4 +1,4 @@
-package racing.utils
+package racing.domain.utils
 
 import kotlin.random.Random
